Rework the tests to make inference with multi-hots work
A lot of the standard ways of doing prediction/inference with Tensorflow don't seem to work with our DenseLayer and/or multi-hot categoricals, but this way seems to. Unfortunately, it requires feeding tensors of different shape at inference time, but it's the only way I could figure out to make this go.