Create `FeatureCollection` as a way to pass input features to blocks
`FeatureCollection` combines schemas and tensor data, providing a similar API to the existing `Schema` class from Merlin Core but here combining it with the actual feature values. This creates a container that we can potentially use inside `ModelContext` to provide the values and schemas of input features to any block in the model without having to pass them through the entire model architecture.

Co-authored-by: Julio Perez <37191411+jperez999@users.noreply.github.com>