Make `Model` creation explicit
Currently, we're assuming that if you give us something that computes a loss, we should create a `Model`, which is a bit of black magic that has some weird edge cases. This change brings the API closer to standard Keras.

We're also looking forward to future changes that make the `Model` responsible for computing losses and metrics, which will break this somewhat magical functionality.

Co-authored-by: Marc Romeyn <marcromeyn@gmail.com>