Remove `Block.to_model`
Previously the Block class had a `to_model` method that would convert the
block to a Model, with an InputBlock and a PredictionTask. This created
a circular dependency between the Block class and the Model class.

Remove this method, in favour of `Model.from_block` . This removes the
dependency on the Model class from Block class - and the block file
no longer needs any imports from Model or RetrievalModel.